backend/requirements.txt
annotated-types==0.7.0
anyio==4.6.0
blinker==1.4
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
coverage==7.6.8
cryptography==3.4.8
dbus-python==1.2.18
distro==1.7.0
distro-info==1.1+ubuntu0.2
dnspython==2.6.1
email_validator==2.2.0
fastapi==0.115.0
fastapi-cli==0.0.5
greenlet==3.1.1
h11==0.14.0
honcho==1.1.0
httpcore==1.0.6
httplib2==0.20.2
httptools==0.6.1
httpx==0.27.2
idna==3.10
importlib-metadata==4.6.4
iniconfig==2.0.0
jeepney==0.7.1
Jinja2==3.1.4
keyring==23.5.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
more-itertools==8.10.0
oauthlib==3.2.0
packaging==24.1
passlib==1.7.4
pluggy==1.5.0
polars==1.9.0
pycairo==1.27.0
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
PyGObject==3.42.1
PyJWT==2.3.0
pyparsing==2.4.7
pytest==8.3.3
python-apt==2.4.0+ubuntu4
python-dotenv==1.0.1
python-multipart==0.0.12
PyYAML==6.0.2
requests==2.32.3
rich==13.9.1
SecretStorage==3.3.1
setuptools==75.1.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.35
sqlmodel==0.0.22
starlette==0.38.6
typer==0.12.5
typing_extensions==4.12.2
unattended-upgrades==0.1
urllib3==2.2.3
uvicorn==0.31.0
uvloop==0.20.0
wadllib==1.3.6
watchfiles==0.24.0
websockets==13.1
zipp==1.0.0
psycopg2-binary==2.9.9